Interface Descriptor changes for module size and lock
Change the interface descriptor field to show the size of the module
that this interface is contained in, and increase the feature mask to 16
bits and add the "this interface has a lock" feature bit.

This puts back in the (pad) field for the interface descriptor as we
can't reuse it for anything, it is part of the "descriptor header".

Acked-by: Karim Yaghmour <karim.yaghmour@opersys.com>
Reviewed-by: Alex Elder <elder@linaro.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>